Have to make this even more generic because we don't have a solid base
Response messages are completely context dependent. means this class is a bit shit now but it still does what I need it to and make things slightly easier to implement. May need to recreate this design at a later time but it seems to do what I want. Also updating the event interface to match